+++ calculation/src/main/java/org/cardanofoundation/rewards/calculation/domain/PoolUpdate.java
@@ -2,6 +2,7 @@ package org.cardanofoundation.rewards.calculation.domain;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.List;
@@ -16,7 +17,7 @@ public class PoolUpdate {
     private String poolId;
     private Integer activeEpoch;
     private Integer retiringEpoch;
-    private Double margin;
+    private BigDecimal margin;
     private BigInteger fixedCost;
     private BigInteger pledge;
     private String rewardAddress;
